Adding default text pages, design update deletion, logging and bug fixing
Design updates
- Fixed bug where the image upload section of design updates wasn't working
- Added logging for items and currency spent on design updates
- Added a check for whether a character is in a trade or transfer before creating a design update
- Added deletion of a design update

Text pages
- Added a script to generate some default text pages with placeholder content (about page, TOS, privacy policy)

Misc
- Added a specification for the table engine for table creation since I had a run-in with it recently
- TinyMCE was forcing locally-uploaded image URLs to be relative, fixed
- Comments

+++ app/Services/PageService.php
@@ -54,6 +54,8 @@ class PageService extends Service
         DB::beginTransaction();
 
         try {
+            if(Config::get('lorekeeper.text_pages.').$page->key) throw new \Exception("You cannot delete this page.");
+
             $page->delete();
 
             return $this->commitReturn(true);
